Implementa segurança com Spring Security: autenticação, autorização por roles ADMIN/USER, endpoints de busca e edição com restrição, ID único em transações

+++ src/main/java/com/example/api/application/repository/TransacaoRepository.java
@@ -17,10 +17,11 @@ public class TransacaoRepository {
     private static long contadorId = 1;
 
     public void adicionarTransacao(Transacao transacao) {
-        transacao.setId(contadorId++);
+        transacao.setId((int) contadorId++);
         lista.add(transacao);
     }
 
+    
     public List<Transacao> listarTransacoes() {
         return new ArrayList<>(lista);
     }
